remove file constants.h, move code to print.h
