+++ src/solver/variable/include/antares/solver/variable/state.h
@@ -32,8 +32,6 @@
 #include "antares/solver/simulation/sim_structure_donnees.h"
 #include "antares/solver/simulation/sim_structure_probleme_economique.h"
 
-#include "constants.h"
-
 namespace Antares::Solver::Variable
 {
 class ThermalState
